+++ src/code/src/basics.rs
@@ -1748,140 +1748,6 @@ fn main() {
 }
 
 #[allow(dead_code)]
-mod app10 {
-use bevy::prelude::*;
-
-    fn server_session() {}
-    fn fetch_server_updates() {}
-    fn keyboard_input() {}
-    fn gamepad_input() {}
-    fn host_debug() {}
-    fn host_session() {}
-    fn host_player_movement() {}
-    fn host_enemy_ai() {}
-    fn smoke_particles() {}
-    fn water_animation() {}
-
-    #[derive(Resource)]
-    struct MyNetworkSession;
-
-    impl MyNetworkSession {
-        fn is_connected(&self) -> bool {
-            true
-        }
-    }
-
-// ANCHOR: run-criteria
-use bevy::ecs::schedule::ShouldRun;
-
-#[derive(Debug, PartialEq, Eq)]
-#[derive(Resource)]
-enum MultiplayerKind {
-    Client,
-    Host,
-    Local,
-}
-
-fn run_if_connected(
-    mode: Res<MultiplayerKind>,
-    session: Res<MyNetworkSession>,
-) -> ShouldRun
-{
-    if *mode == MultiplayerKind::Client && session.is_connected() {
-        ShouldRun::Yes
-    } else {
-        ShouldRun::No
-    }
-}
-
-fn run_if_host(
-    mode: Res<MultiplayerKind>,
-) -> ShouldRun
-{
-    if *mode == MultiplayerKind::Host || *mode == MultiplayerKind::Local {
-        ShouldRun::Yes
-    } else {
-        ShouldRun::No
-    }
-}
-
-fn main() {
-    App::new()
-        .add_plugins(DefaultPlugins)
-
-        // if we are currently connected to a server,
-        // activate our client systems
-        .add_system_set(
-            SystemSet::new()
-                .with_run_criteria(run_if_connected)
-                .before("input")
-                .with_system(server_session)
-                .with_system(fetch_server_updates)
-        )
-
-        // if we are hosting the game,
-        // activate our game hosting systems
-        .add_system_set(
-            SystemSet::new()
-                .with_run_criteria(run_if_host)
-                .before("input")
-                .with_system(host_session)
-                .with_system(host_player_movement)
-                .with_system(host_enemy_ai)
-        )
-
-        // other systems in our game
-        .add_system(smoke_particles)
-        .add_system(water_animation)
-        .add_system_set(
-            SystemSet::new()
-                .label("input")
-                .with_system(keyboard_input)
-                .with_system(gamepad_input)
-        )
-        .run();
-}
-// ANCHOR_END: run-criteria
-
-mod sub {
-    use super::*;
-
-// ANCHOR: run-criteria-label
-#[derive(Debug, Clone, PartialEq, Eq, Hash)]
-#[derive(RunCriteriaLabel)]
-enum MyRunCriteria {
-    Client,
-    Host,
-}
-
-fn main() {
-    App::new()
-        // ...
-        .add_system_set(
-            SystemSet::new()
-                .with_run_criteria(
-                    // assign it a label
-                    run_if_host
-                        .label(MyRunCriteria::Host)
-                )
-                .with_system(host_session)
-                .with_system(host_player_movement)
-                .with_system(host_enemy_ai)
-        )
-
-        // extra system for debugging the host
-        // it can share our previously-registered run criteria
-        .add_system(host_debug
-            .with_run_criteria(MyRunCriteria::Host)
-        )
-        .run();
-}
-
-// ANCHOR_END: run-criteria-label
-}
-}
-
-#[allow(dead_code)]
 mod app11 {
     use bevy::prelude::*;
     use super::*;
